% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perr_cox.R
\name{perr_cox}
\alias{perr_cox}
\title{PERR method with Cox proportional hazards regression}
\usage{
perr_cox(
  data,
  treatment_var,
  time_var,
  event_var,
  covariates = NULL,
  interactions = NULL,
  nonlinear_terms = NULL
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{treatment_var}{The name of the treatment variable in the data.}

\item{covariates}{A character vector of covariate names. Default is NULL.}

\item{interactions}{A character vector of interaction term names. Default is NULL.}

\item{nonlinear_terms}{A character vector of nonlinear term names. Default is NULL.}

\item{outcome_var}{The name of the time-to-event outcome variable in the data.}
}
\value{
A list containing the original and adjusted hazard ratios (HR) with their 95\% confidence intervals.
}
\description{
This function performs the PERR method using Cox proportional hazards regression to adjust the treatment effect estimate.
}
\examples{
\dontrun{
data <- data.frame(treatment = c(0,1,0,1), outcome = c(5,3,2,4))
perr_cox(data, "treatment", "outcome")
}
}
