% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perr_linear.R
\name{perr_linear}
\alias{perr_linear}
\title{PERR method with linear regression}
\usage{
perr_linear(
  data,
  treatment_var,
  outcome_var,
  covariates = NULL,
  interactions = NULL,
  nonlinear_terms = NULL
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{treatment_var}{The name of the treatment variable in the data.}

\item{outcome_var}{The name of the continuous outcome variable in the data.}

\item{covariates}{A character vector of covariate names. Default is NULL.}

\item{interactions}{A character vector of interaction term names. Default is NULL.}

\item{nonlinear_terms}{A character vector of nonlinear term names. Default is NULL.}
}
\value{
A list containing the original and adjusted beta coefficients with their 95\% confidence intervals.
}
\description{
This function performs the PERR method using linear regression to adjust the treatment effect estimate.
}
\examples{
\dontrun{
data <- data.frame(treatment = c(0,1,0,1), outcome = c(5.2,3.4,2.1,4.5))
perr_linear(data, "treatment", "outcome")
}
}
