% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perr_logistic.R
\name{perr_logistic}
\alias{perr_logistic}
\title{PERR method with logistic regression}
\usage{
perr_logistic(
  data,
  treatment_var,
  outcome_var,
  covariates = NULL,
  interactions = NULL,
  nonlinear_terms = NULL
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{treatment_var}{The name of the treatment variable in the data.}

\item{outcome_var}{The name of the binary outcome variable in the data.}

\item{covariates}{A character vector of covariate names. Default is NULL.}

\item{interactions}{A character vector of interaction term names. Default is NULL.}

\item{nonlinear_terms}{A character vector of nonlinear term names. Default is NULL.}
}
\value{
A list containing the original and adjusted odds ratios (OR) with their 95\% confidence intervals.
}
\description{
This function performs the PERR method using logistic regression to adjust the treatment effect estimate.
}
\examples{
\dontrun{
data <- data.frame(treatment = c(0,1,0,1), outcome = c(1,0,1,1))
perr_logistic(data, "treatment", "outcome")
}
}
